+++ packages/changed-elements-react/src/NamedVersionSelector/TextEx.tsx
@@ -6,7 +6,7 @@ import { type ComponentProps, type CSSProperties } from "react";
 import { Text } from "@itwin/itwinui-react";
 import clsx from "clsx";
 
-import "./TextEx.css";
+import "./TextEx.scss";
 
 interface TextExProps extends ComponentProps<typeof Text> {
   weight?: "light" | "normal" | "semibold" | "bold" | undefined;
@@ -46,11 +46,13 @@ export function TextEx(props: TextExProps): ReturnType<typeof Text> {
     overflow,
     color,
     children,
+    variant,
     ...rest
   } = props;
   return (
     <Text
       {...rest}
+      data-variant={variant}
       as={oblique ? "i" : "div"}
       className={clsx("_cer_v1_text", className)}
       data-weight={weight}