+++ packages/changed-elements-react/src/NamedVersionSelector/useNamedVersionsList.ts
@@ -140,16 +140,7 @@ export function useNamedVersionsList(args: UseNamedVersionListArgs): UseNamedVer
           if (currentChangesetArrayIndex === -1) {
             setIsLoading(false);
             setIsError(true);
-            setCurrentNamedVersion({
-              id: "error",
-              displayName: IModelApp.localization.getLocalizedString(
-                "VersionCompare:versionCompare.namedVersionErrorPlaceholder",
-              ),
-              changesetId: currentChangesetId,
-              changesetIndex: -1,
-              description: null,
-              createdDateTime: "",
-            });
+            setCurrentNamedVersion(undefined);
             return;
           }